wrote tests created and changed some classes

+++ src/com/mishamba/day1/model/Oven.java
@@ -5,7 +5,7 @@ public class Oven extends HouseholdAppliance {
     private int depth;
 
     public Oven(int powerConsumption, int weight, int capacity, int depth,
-                int height, double width) {
+                double height, double width) {
         super(height, weight, width ,powerConsumption);
         this.capacity = capacity;
         this.depth = depth;
